feat: 保存游戏数据

+++ src/context.zig
@@ -2,4 +2,4 @@ const std = @import("std");
 const zhu = @import("zhu");
 
 pub var battleNpcIndex: u16 = 0;
-pub var oldMapIndex: u16 = 0;
+pub var oldMapIndex: u8 = 0;
